Mango distribution
Get the user input of Integer list N - number of mangoes in a bag and M - number of students. Then distribute the mangoes to each student and make sure each student gets one bag. Find the difference between the number of mangoes in a bag maximum mangoes and bag with a minimum number of mangoes given to the student is minimum.